pyproject.toml
[tool.poetry]
name = "kauri-test"
version = "0.1.0"
description = ""
authors = ["Mark R <theiterance@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1"
pydantic = "^2.8.2"
channels = {extras = ["daphne"], version = "^4.1.0"}
websockets = "^13.0.1"
loguru = "^0.7.2"
httpx = "^0.27.2"
django-redis = "^5.4.0"
pydantic-settings = "^2.4.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
